backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
databases
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv